Start sharing code between pd_shift and insert
